plugins.svtplay: add subtitle language metadata for ffmpeg
Grabs the language of the subtitle from the DASH MPD to use instead of
the format name "webvtt" as subtitle metadata for muxing with ffmpeg.

I'm not sure SVT Play offer subtitles in any language other than
Swedish. There are some episodes with complete burn-ins, but they are
rare exceptions. I've been in contact with them about this, and they
told me translations to other languages are not something they put a
lot of resources on at this time, but that they might offer more of it
in the future as technology improves and new shiny things come to
offer a viable solution to this type of accessibility.

So with this in mind, checking the MPD might seem like overkill but at
least no one's hands are tied if they ever where encounter a subtitle in
any other language from SVT Play.